fixed Decimal mistake in stock_item and movement

+++ backend/app/crud/stock_movement.py
@@ -35,7 +35,7 @@ class CRUDStockMovement(CRUDBase[StockMovement, StockMovementCreate, StockMoveme
         *,
         stock_item: StockItem,
         type: str,
-        quantity_delta: float,
+        quantity_delta: Decimal,
         description: str | None = None,
         prevent_negative: bool = True,
     ) -> StockMovement:
